Moved appointment line item to more relevant folder structure
