+++ src/Components/AsyncLoadingWrapper.js
@@ -1,27 +1,28 @@
-import { Box } from "@material-ui/core";
+import { Box, CircularProgress } from "@material-ui/core";
 import React from "react";
-import Loading from "../Practitioner/Shared/Loading";
 
 export default function AsyncLoadingWrapper({ status, children }) {
   switch (status) {
+    case "idle":
+        <></>
     case "pending":
       return (
         <Box
-          minHeight="100px"
+          minHeight="300px"
           display="flex"
           alignItems="center"
           justifyContent="center"
         >
-          <Loading />
+          <CircularProgress variant="indeterminate" />
         </Box>
       );
     case "success":
-        return <>{children}</>
-    
+      return <>{children}</>;
+
     case "error":
-        return <p>Error with network request, please try again</p>
+      return <p>Error with network request, please try again</p>;
 
     default:
-      return <p>Error, unexpected state</p>
+      return <p>Error, unexpected state</p>;
   }
 }
